recipes-core: datapartition: change the way to udpate a device
We now concentrate on the SD card update, and remove the support for OTA
updates. the OTA update scheme was incomplete and never used. Instead,
we support only update of the SD card, and /var/lib/docker + /home/root
can be optionnally mounted to the extra data partition (/data) that is
dynamically built, filling up the extra space on the provided SD card.